Fix circular dependency between the 2 vault objects.
